fix(database): resolve decimal type mismatch and improve connection handling
- Fix type mismatch between PostgreSQL DECIMAL and Rust types
  - Replace f64 with rust_decimal::Decimal for price fields
  - Add rust_decimal and rust_decimal_macros dependencies
  - Update SQLx features to include rust_decimal support

- Update product entity and DTOs
  - Change price field type from f64 to Decimal
  - Update validation logic to use dec! macro
  - Ensure type consistency across the application

- Improve database connection handling
  - Add proper error logging
  - Add connection pool configuration
  - Add connection verification

Technical details:
- Use exact decimal arithmetic for monetary values
- Prevent floating-point precision issues
- Maintain type safety with PostgreSQL
- Fix products list endpoint functionality

Dependencies added:
- rust_decimal = 1.32
- rust_decimal_macros = 1.32

+++ src/interfaces/http/responses/product_responses.rs
@@ -2,13 +2,14 @@ use serde::Serialize;
 use uuid::Uuid;
 use chrono::{DateTime, Utc};
 use crate::domain::entities::product::Product;
+use rust_decimal::Decimal;
 
 #[derive(Debug, Serialize)]
 pub struct ProductResponse {
     pub id: Uuid,
     pub name: String,
     pub description: String,
-    pub price: f64,
+    pub price: Decimal,
     pub stock: i32,
     pub created_at: DateTime<Utc>,
     pub updated_at: DateTime<Utc>,
